@LucasWilkinson 2025-02-11
@khluu can we run the perf CI on this? would be nice to check for regressions since theres alot of FA changes