@LucasWilkinson 2025-03-07
not seeing much speedup (minor TFTT)

```
8xH200, DeepSeek-R1, VLLM_USE_V1=1 VLLM_ATTENTION_BACKEND=FLASHMLA VLLM_USE_FLASHINFER_SAMPLER=1

Main:

  backend  input_tokens  output_tokens  output_toks/s     req/s  median_itl_ms  median_ttft_ms
2    vllm          1000           1000    1095.323697  1.095324      40.931626      149.658605
1    vllm          5000           1000     517.327850  0.517328      39.956240     5627.535715
3    vllm         10000           1000     315.639817  0.315640      39.697455    57821.907031
0    vllm         32000           1000     106.821047  0.106821      40.109005   193232.262791


This PR:

  backend  input_tokens  output_tokens  output_toks/s     req/s  median_itl_ms  median_ttft_ms
2    vllm          1000           1000    1034.089425  1.034089      39.884763     2365.846871
1    vllm          5000           1000     518.861897  0.518862      39.906694     5312.827895
3    vllm         10000           1000     317.878565  0.317879      40.045355    57160.958344
0    vllm         32000           1000     110.514999  0.110515      40.538517   184969.578100
```