@chaunceyjiang 2025-04-18
Hi, @LucasWilkinson @simon-mo. This PR uses a new function `get_scheduler_metadata`:

https://github.com/vllm-project/flash-attention/blob/main/vllm_flash_attn/flash_attn_interface.py#L78

It seems that a new release of `vllm_flash_attn` is needed.