@chaunceyjiang 2025-04-18
https://pypi.org/project/vllm-flash-attn/

It’s been a long time since `vllm-flash-attn` had a new release. Should we consider publishing a new version?